sensor_cloud_security/requirements.txt
Django==1.10.5
django-mysql==1.1.0
djangorestframework==3.5.4
mysqlclient==1.3.9
Pygments==2.2.0
django-oauth-toolkit==0.11.0
purl==1.3
requests==2.13.0
django-cors-middleware==1.3.1